Polish Alternating Flagstones tutorial; fix pre_conway leak and dtype mismatch
Notebook (docs/notebooks/Alternating Flagstones.ipynb):
- Drop matplotlib subplot rendering of half-edge graphs; use multi_show
  everywhere so panel titles and Cairo output are consistent with the
  rest of the curated series.
- Use the {7,3} dual + Euclidean projection + normalize_edge_lengths()
  as the running example from section 2 onward (was platonic squares
  for sections 1-2 then dropped).
- Drop the manual 'pre_conway' workaround in the {7,3} cell — fixed in
  build_structure (see below).
- Split section 4 (boundary cleanup, before/after extend_border) from
  section 5 (folded preview, which now bundles cut_twist_centres +
  fold_wireframe). Boundary cleanup is purely visual; flat-folding
  needs the cut.
- Display the optimisation before/after side-by-side.
- Subdivision section shows straight vs. curved-fold ridges side-by-side.

Library:
- alternating_flagstones.build_structure now strips any incoming
  'pre_conway' attributes from its working copy before applying the
  Conway operator, so chaining (e.g. dual_graph -> alternating_*)
  works without manual cleanup.
- alternating_flagstones.optimize_alternating_flagstone writes back
  float64 vertex positions (PyTorch defaults to float32, which broke
  downstream numba code).
- cutting.cut_out_poly now casts both segment arrays to float64 before
  passing them to the JIT'd line_segment_intersections — fixes the
  TypingError that surfaced when cutting twist centres on an
  optimised CP.
- rendering.CairoRenderer.auto_line_width falls back to computing
  edge lengths on the fly when 'length' is not cached on halfedges,
  unbreaking 5 rendering tests against ungeometried graphs.

Tests:
- New scaling-helper tests in tests/test_half.py: TestScalePositions
  (multiplies positions, recomputes lengths, rejects non-Euclidean)
  and TestNormalizeEdgeLengths (geometric / arithmetic / factor /
  invalid mode).

343 fast tests pass; mkdocs --strict builds cleanly.

+++ eucare/cutting.py
@@ -185,9 +185,12 @@ def delete_new_outside(G: "HalfEdgeGraph", border_key: str = 'new_border') -> No
 
 
 def cut_out_poly(G: "EuclideanPositionHEG", poly, delete_outside: bool = True, eps: float = 1e-10):
-    poly_segments = np.stack(list(rotate_by(poly, (0, 1))))
+    # numba-jit'd ``line_segment_intersections`` requires a uniform float64
+    # dtype across both segment arrays; cast defensively here so that callers
+    # that produced positions via float32 paths (e.g. torch optimisers) work.
+    poly_segments = np.stack(list(rotate_by(poly, (0, 1)))).astype(np.float64)
     es = list(G.halfedges_representing_edges())
-    edge_segments = np.array([[e.orig['pos'], e.dest['pos']] for e in es])
+    edge_segments = np.array([[e.orig['pos'], e.dest['pos']] for e in es], dtype=np.float64)
     crossing_positions, segments1_to_crossings, segments2_to_crossings, crossings_to_segments1, crossings_to_segments2 = \
         get_ordered_crossings(poly_segments, edge_segments, eps=eps)
